latentmeter/requirements.txt
python-dotenv
aiohttp
pyTelegramBotAPI # https://github.com/eternnoir/pyTelegramBotAPI#asynchronous-delivery-of-messages